requirements.txt
python-telegram-bot==20.7
gspread
google-auth
python-certifi-win32